Differentate cached real vs fake data
We have only one slot for cached data. Don't serve cached fake data when
asked for real data, and vice versa.

We're keeping the caching mechanism for fake data because we want the
serving path to be as similar to real data as possible. The idea is that
fake data is a generalized testing mechanism.